Move Google login to a subapp
Move /login_google and /token_google endpoints to a subapp, so only
these 2 endpoints use the Session middleware, instead of the entire
app. FastAPI doesn't support router middlewares, so a subapp is
required.
