+++ app/main.py
@@ -1,9 +1,8 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from starlette.middleware.sessions import SessionMiddleware
 
 from app.api.api import api_router
-from app.core.config import get_settings
+from app.api.endpoints.login_google import google_auth_app
 from app.database.db import Base, engine
 
 # create DB tables (would be better to use Alembic)
@@ -21,7 +20,8 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# add Session middleware so Authlib can access the request session
-app.add_middleware(SessionMiddleware, secret_key=get_settings().SECRET_KEY)
-
 app.include_router(api_router, prefix="/api")
+
+# use a subapp so the Session middleware required for Google authentication
+# is only present in those endpoints
+app.mount("/api", google_auth_app)
