+++ app/tests/factories/__init__.py
@@ -1,2 +1,4 @@
 from .todo_factory import ToDoFactory
 from .user_factory import UserFactory
+
+factory_list = [ToDoFactory, UserFactory]
